#12 Feat: New Views added in Service mode

+++ AT2/src/main/java/com/dpo/centralized_restaurant/database/ConectorDB.java
@@ -315,7 +315,7 @@ import java.util.ArrayList;
                             rs.getString(0),
                             rs.getDouble(1),
                             rs.getInt(2),
-                            rs.getDouble(3)
+                            rs.getInt(3)
                     );
                     result.add(aux);
                 }
@@ -345,7 +345,7 @@ import java.util.ArrayList;
                             rs.getString("name"),
                             rs.getDouble("cost"),
                             rs.getInt("units"),
-                            rs.getDouble("timecost")
+                            rs.getInt("timecost")
                     );
                     result.add(aux);
                 }
@@ -391,7 +391,7 @@ import java.util.ArrayList;
                             rs.getString(0),
                             rs.getDouble(1),
                             rs.getInt(2),
-                            rs.getDouble(3)
+                            rs.getInt(3)
                     );
                     result.add(aux);
                 }
